split apps/cli/src/core/platform-resolver.ts into     - project-discovery.ts - Finding project root and paths     - environment-loader.ts - Loading and merging environment configurations     - service-resolver.ts - Resolving service platform assignments     - platform-types.ts - Platform type definitions     - configuration-error.ts - Custom error class
