+++ apps/cli/src/__tests__/environment-validation.test.ts
@@ -9,7 +9,7 @@ import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as os from 'os';
-import { getAvailableEnvironments, isValidEnvironment, loadEnvironmentConfig } from '../core/platform-resolver.js';
+import { getAvailableEnvironments, isValidEnvironment, loadEnvironmentConfig } from '../core/environment-loader.js';
 
 describe('Dynamic Environment Validation', () => {
   let testDir: string;
