+++ apps/cli/src/core/command-results.ts
@@ -5,7 +5,7 @@
  * commands operate on multiple services or need to track overall execution.
  */
 
-import type { PlatformType } from './platform-resolver.js';
+import type { PlatformType } from './platform-types.js';
 
 // Base result interface that all command results extend (for testing/mocking)
 export interface BaseCommandResult {
