+++ apps/cli/src/core/command-service-matcher.ts
@@ -21,7 +21,8 @@
 
 import { commandRequiresServices } from './command-discovery.js';
 import { getAvailableServices, isValidService, ServiceSelector, ServiceCapability, ServiceName } from './service-discovery.js';
-import { findProjectRoot, resolveServiceDeployments } from './platform-resolver.js';
+import { findProjectRoot } from './project-discovery.js';
+import { resolveServiceDeployments } from './service-resolver.js';
 import { ServiceFactory } from '../services/service-factory.js';
 import { parseEnvironment } from './environment-validator.js';
 import { serviceSupportsCommand } from './service-command-capabilities.js';
