+++ apps/cli/src/core/dashboard/log-aggregator.ts
@@ -2,7 +2,7 @@
  * Log Aggregator - Fetches and aggregates logs from various sources
  */
 
-import { type ServicePlatformInfo } from '../platform-resolver.js';
+import { type ServicePlatformInfo } from '../service-resolver.js';
 
 export interface LogEntry {
   timestamp: Date;
