+++ apps/cli/src/core/dashboard/watch-dashboard.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import { render } from 'ink';
 import { DashboardApp, DashboardMode } from '../dashboard/dashboard-layouts.js';
 
-import { getAvailableEnvironments } from '../platform-resolver.js';
+import { getAvailableEnvironments } from '../environment-loader.js';
 
 // Argument parsing with environment support
 function parseArgs(): { environment: string; mode: DashboardMode } {
