+++ apps/cli/src/core/state-manager.ts
@@ -8,7 +8,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import { ServiceName } from './service-discovery.js';
-import { PlatformType } from './platform-resolver.js';
+import { PlatformType } from './platform-types.js';
 import { PlatformResources } from '../platforms/platform-resources.js';
 
 export interface ServiceState {
