+++ apps/cli/src/platforms/aws/cloudwatch-fetcher.ts
@@ -12,7 +12,7 @@ import {
   DescribeStacksCommand 
 } from '@aws-sdk/client-cloudformation';
 import { LogFetcher, LogEntry, LogFetchOptions } from '../../core/dashboard/log-aggregator.js';
-import { type ServicePlatformInfo } from '../../core/platform-resolver.js';
+import { type ServicePlatformInfo } from '../../core/service-resolver.js';
 
 export class CloudWatchLogFetcher extends LogFetcher {
   private logsClient: CloudWatchLogsClient;
