+++ apps/cli/src/platforms/aws/handlers/ecs-publish.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 import * as fs from 'fs';
 import { AWSPublishHandlerContext, PublishHandlerResult, HandlerDescriptor } from './types.js';
 import { printInfo, printSuccess } from '../../../core/io/cli-logger.js';
-import { loadEnvironmentConfig } from '../../../core/platform-resolver.js';
+import { loadEnvironmentConfig } from '../../../core/environment-loader.js';
 
 /**
  * Publish handler for ECS Fargate services
