+++ apps/cli/src/platforms/aws/platform.ts
@@ -67,7 +67,7 @@ export class AWSPlatform extends Platform {
     appStack?: string;
   } {
     // Load the environment configuration to get AWS settings
-    const { loadEnvironmentConfig } = require('../../core/platform-resolver.js');
+    const { loadEnvironmentConfig } = require('../../core/environment-loader.js');
     const envConfig = loadEnvironmentConfig(service.environment);
     
     // Get AWS config from environment file, fallback to env vars
